server/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
pdf2image==1.16.3
Pillow==10.3.0
